third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-invalid-args-in-workers-expected.txt

[Worker] Test createImageBitmap with invalid arguments in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: ./resources/canvas-createImageBitmap-invalid-args-in-workers.js
PASS [Worker] createImageBitmap(data, 0, 0, 10, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS [Worker] createImageBitmap(data, 0, 0, 0, 10) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS successfullyParsed is true

TEST COMPLETE
